# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for CampusPulse college events and participation.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorDetails = { [key: string]: unknown };

export interface Error {
  error: string;
  details?: ErrorDetails;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  ADMIN: 'ADMIN',
  STUDENT: 'STUDENT',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  created_at: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RefreshInput {
  refresh: string;
}

export interface TokenResponse {
  access: string;
  refresh?: string;
}

export type AuthResponse = TokenResponse & {
  user: User;
};

export interface Category {
  id: number;
  name: string;
  description: string;
  created_at: string;
  event_count: number;
}

export interface CategoryInput {
  /** @minLength 2 */
  name: string;
  description?: string;
}

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


export const EventStatus = {
  UPCOMING: 'UPCOMING',
  ONGOING: 'ONGOING',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;

export type EventCategory = {
  id: number;
  name: string;
};

export interface Event {
  id: number;
  title: string;
  description: string;
  category: EventCategory;
  date: string;
  start_time: string;
  end_time: string;
  venue: string;
  organizer: string;
  capacity: number;
  registration_deadline: string;
  status: EventStatus;
  created_at: string;
  updated_at: string;
  participant_count: number;
}

export interface EventInput {
  /** @minLength 2 */
  title: string;
  /** @minLength 1 */
  description: string;
  category: number;
  date: string;
  start_time: string;
  end_time: string;
  venue: string;
  organizer: string;
  /** @minimum 1 */
  capacity: number;
  registration_deadline: string;
  status?: EventStatus;
}

export type RegistrationStatus = typeof RegistrationStatus[keyof typeof RegistrationStatus];


export const RegistrationStatus = {
  REGISTERED: 'REGISTERED',
  CANCELLED: 'CANCELLED',
  ATTENDED: 'ATTENDED',
  ABSENT: 'ABSENT',
} as const;

export interface Registration {
  id: number;
  student: User;
  event: Event;
  registration_date: string;
  status: RegistrationStatus;
}

export interface RegistrationUpdate {
  status: RegistrationStatus;
}

export interface DashboardStats {
  total_events: number;
  upcoming_events: number;
  completed_events: number;
  cancelled_events: number;
  total_students: number;
  total_registrations: number;
  recent_events: Event[];
  my_registrations: number;
  available_events: number;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Permission denied
 */
export type ForbiddenResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type ListEventsParams = {
search?: string;
category?: number;
status?: EventStatus;
date?: string;
ordering?: ListEventsOrdering;
};

export type ListEventsOrdering = typeof ListEventsOrdering[keyof typeof ListEventsOrdering];


export const ListEventsOrdering = {
  date: 'date',
  '-date': '-date',
  title: 'title',
  '-title': '-title',
  '-created_at': '-created_at',
} as const;