# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: REST API for CampusPulse college events and participation.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/refresh/:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: Refreshed token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details:
          type: object
          additionalProperties: true
...
    RefreshInput:
      type: object
      required: [refresh]
      properties:
        refresh: { type: string }
    TokenResponse:
      type: object
      required: [access]
      properties:
        access: { type: string }
        refresh: { type: string }
...