# lib/api-zod/src/generated/types/error.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for CampusPulse college events and participation.
 * OpenAPI spec version: 0.1.0
 */
import type { ErrorDetails } from './errorDetails';

export interface Error {
  error: string;
  details?: ErrorDetails;
}